refactor(replies): extract useRegisterFreshReplies hook per CodeRabbit
