+++ src/components/post-mobile/post-mobile.tsx
@@ -36,7 +36,7 @@ import useReplyModalStore from '../../stores/use-reply-modal-store';
 import { selectPostMenuProps } from '../../lib/utils/post-menu-props';
 import useChallengesStore from '../../stores/use-challenges-store';
 import useFeedResetStore from '../../stores/use-feed-reset-store';
-import usePostNumberStore from '../../stores/use-post-number-store';
+import useRegisterFreshReplies from '../../hooks/use-register-fresh-replies';
 import { alertChallengeVerificationFailed } from '../../lib/utils/challenge-utils';
 import useQuotedByMap from '../../hooks/use-quoted-by-map';
 import useProgressiveRender from '../../hooks/use-progressive-render';
@@ -610,6 +610,7 @@ const PostMobile = ({
   const updatedReplies = (repliesResult as { updatedReplies?: Comment[] }).updatedReplies;
   const repliesForRender = updatedReplies?.length ? updatedReplies : replies || [];
   const freshRepliesForRender = useFreshReplies(repliesForRender);
+  useRegisterFreshReplies(post, freshRepliesForRender);
   const reset = (repliesResult as { reset?: () => Promise<void> }).reset;
   const setResetFunction = useFeedResetStore((s) => s.setResetFunction);
   useEffect(() => {
@@ -619,22 +620,6 @@ const PostMobile = ({
       });
     }
   }, [isInPostView, isInPendingPostView, reset, setResetFunction]);
-  const registerComments = usePostNumberStore((s) => s.registerComments);
-  const prevCidsRef = useRef<string>('');
-  useEffect(() => {
-    const all = post ? [post, ...freshRepliesForRender] : freshRepliesForRender;
-    if (!all.length) return;
-    const cidsKey = all
-      .map((comment) => {
-        const commentKey = comment?.cid ?? (typeof comment?.index === 'number' ? `index:${comment.index}` : `timestamp:${comment?.timestamp ?? ''}`);
-        return `${comment?.subplebbitAddress ?? ''}:${commentKey}:${typeof comment?.number === 'number' ? comment.number : ''}`;
-      })
-      .sort()
-      .join(',');
-    if (cidsKey === prevCidsRef.current) return;
-    prevCidsRef.current = cidsKey;
-    registerComments(all);
-  }, [post, freshRepliesForRender, registerComments]);
 
   const isInPostPageView = isPostPageView(location.pathname, params);
   const { hidden, unhide } = useHide({ cid });
